requirements.txt
# Requirements per Magazzino WebApp
# Generato automaticamente

# Framework Web
Flask==3.0.0
Werkzeug==3.0.1

# Database
mysql-connector-python==9.1.0

# Elaborazione Dati
pandas>=2.2.0
numpy>=1.26.0

# Excel/File Processing
openpyxl==3.1.2
XlsxWriter==3.1.9

# Utilità
python-dateutil==2.8.2
python-dotenv==1.0.0
